Refactor project structure
Moved the database schema to it's own file in the models directory.
Created a utils folder that contains config.js which holds the env variables
such as the MongoDB URI and PORT. Separated index.js and app.js logic.
index.js now imports app.js since it is the entry point for the backend. This means that app.js now has to be exported and we need to use
http to create the server since we don't import express into index.js