Complete exercise 4.5
Removed the dummy function and test. Added a function to
find the blog with the most likes in a blog list and return it.
If 2 blogs have the same number of likes it will return the first of
those blogs encountered in the list. Added tests for this functionality
and refactored the variable declarations in the test file to be global as they
are and will be used multiple times while testing.