Complete exercise 4.6
Added a function to find the author who has written the most amount of blogs
and returns a new object that contains the author and number of blogs
that they have written. Also wrote a test for this functionality that
gives a list of multiple blogs as input to the function.