Complete exercise 4.8
Installed the supertest dependency and created a test file for the
blogs api. The tests use a seperate test database that resets to an intial
state before every test is run to ensure testing consistency. Edited the logger file
to only log information/errors if the NODE_ENV is not test