Complete exercise 4.12
Added express-async-errors as a dependency to keep readability
high since no try-catch blocks. Added an error handler middleware
to handle catch statements depending on the type of error. In this case
the only error being caught is a ValidationError that returns a 400 bad
request status and the error message. Tested that this validation error occurs
when the schema validation is broken in the case of no title or no url.