Complete exercise 4.13
Added delete request functionality to the bloglist application that takes an id request parameter
and deletes the resource located at /api/blogs/:id. Created a test suite for this functionality
that tests 4 things: delete an existing resource, delete a resource already deleted, delete
a resource that has an unidentified id, and delete a resource with a malformatted id.