Add get request error handling and tests
Previously, the get request to a specific id even if that id didn't exist
returned a 200 status code. The logic has been updated so that in this case
a 404 is returned as that resource is not found. Added tests for this functionality
as well as a test to check that getting a resource at an existing id is successful.