Complete exercise 4.15
Added POST request functionality to to the user api endpoint that creates and saves a user to
the database. The user contains fields username, name, and password. Since we don't want to
store the password in text form in the database, we instead leverage the bcrypt library
to salt and hash the passwords and that password hash will then be stored in the database.
Added tests to check that the POST request functionality works.