Complete exercise 4.16
Added restrictions to the username and password fields of the user schema. Now the request
must have a username field and password field with each of their lengths being greater
than or equal to 3 characters. Validation for the password is conducted in the controller
instead of mongoose validation because the password hash and password itself or different things.
Added tests to determine that we recieve the appropriate status code if the username or password does
not pass validation.