Update model schema and existing user check
As preparation for linking a user with the blogs they have posted and vice versa, updated
the model schema to include references to the respective id's of user or blogs. Also added
validation for existing usernames in the database and a test to check that this functionality
works as intended.