Complete exercise 4.17
Updated the model schema for users and blogs so that they contain an array of ObjectId's
instead of one sole ObjectId. Post requests to blogs now find one user to link the
user id to the blog. Then the blog is concatenated to the users blogs array so that
the user has reference to the blogs ObjectId and the blog has reference to the
users ObjectId. Updated the get requests for both blogs and users so that they populate
the array of ObjectId's with their actual content fields.