Complete exercise 4.18
Added token authentication to the bloglist application via the json web token library.
The token is given when the user logs in successfully by making a post request with
the user's username and password to /api/login. The password in the request body
is compared to the password hash for the respective user via bcrypt.compare. If the user
and password exist and are correct, then we sign a json web token given the user name and id
and return it to the browser for future authentication and linking what the user has done.