Complete exercise 4.21
Added token atuhentication to delete requests so that only users who post the blog are allowed to delete it.
If a user attempts to delete a blog they didn't post, a 401 unauthorized is returned. Removing the blog
has also been separated into different steps. Previously, we would just findByIdAndRemove the blog, however
we need to check if the user deleting the blog is the same user who posted it. In this case we first find the blog
and then compare the blog's user id to the id linked with the token. If they are equivalent then we know the user
who created the blog is also the one trying to delete it.