Complete exercise 4.22
Refactored user extraction to a middleware that only functions on post and delete requests.
Essentially when we verify the token in this middleware, jwt.verify returns the user information
linked with the token (in this case username and id). Then we set this value to request.user to be
accessed in the post and delete requests when comparing the user making the request and the user
in the database